services/collector/requirements.txt
Flask[async]==3.1.0
protobuf==5.29.2
opentelemetry-proto==1.29.0